[Week 7] 이벤트 기반 트랜잭션 분리

+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderInfo.java
@@ -38,7 +38,7 @@ public record OrderInfo(
             return new OrderItemInfo(
                     orderItemEntity.getId(),
                     orderItemEntity.getProductId(),
-                    orderItemEntity.getQuantity(),
+                    orderItemEntity.getQuantity().intValue(),
                     orderItemEntity.getPrice()
             );
         }
